+++ src/components/common/vatsim/CommonAirportCard.vue
@@ -216,6 +216,7 @@ const controllers = computed<Controller[]>(() => {
                 -webkit-line-clamp: 2;
 
                 text-align: center;
+                word-break: break-word;
                 white-space: unset;
             }
         }