+++ src/components/map/filters/settings/MapSettingsLayers.vue
@@ -1,5 +1,13 @@
 <template>
     <div class="__info-sections">
+        <common-button
+            v-if="Object.keys(store.mapSettings).length"
+            size="S"
+            type="secondary-875"
+            @click="resetActive = true"
+        >
+            Reset to defaults
+        </common-button>
         <common-block-title>
             General
         </common-block-title>
@@ -9,19 +17,18 @@
         >
             Traffic Heatmap
         </common-toggle>
-        <div class="__section-group">
-            <div class="__small-title">
+        <div class="__grid-info-sections __grid-info-sections--large-title">
+            <div class="__grid-info-sections_title">
                 Aircraft scale
             </div>
-            <input
-                max="1.5"
-                min="0.5"
-                step="0.05"
-                type="range"
-                :value="String(store.mapSettings.aircraftScale ?? 1)"
-                @input="setUserMapSettings({ aircraftScale: Number(($event.target as HTMLInputElement).value) })"
-            >
-            x{{ store.mapSettings.aircraftScale ?? 1 }}
+            <common-select
+                :items="scaleOptions"
+                max-dropdown-height="200px"
+                :model-value="store.mapSettings.aircraftScale ?? null"
+                placeholder="Choose Scale"
+                width="100%"
+                @update:modelValue="setUserMapSettings({ aircraftScale: $event as number })"
+            />
         </div>
         <common-block-title>
             Airports Counters
@@ -79,6 +86,35 @@
                 When <strong>not</strong> disabled, those aircraft are always excluded<br> from dep list when on ground
             </template>
         </common-toggle>
+
+        <common-popup v-model="resetActive">
+            <template #title>
+                Map Settings Reset
+            </template>
+
+            You are about to reset all your map settings to defaults. This action is permanent.
+
+            <template #actions>
+                <common-button
+                    type="secondary-flat"
+                    @click="[resetUserMapSettings(), resetActive = false]"
+                >
+                    Confirm reset
+                </common-button>
+                <common-button
+                    type="secondary-875"
+                    @click="backupMapSettings()"
+                >
+                    Backup data
+                </common-button>
+                <common-button
+                    type="primary"
+                    @click="resetActive = false"
+                >
+                    Cancel that
+                </common-button>
+            </template>
+        </common-popup>
     </div>
 </template>
 
@@ -89,25 +125,30 @@ import type { IUserMapSettings } from '~/utils/backend/map-settings';
 import CommonSelect from '~/components/common/basic/CommonSelect.vue';
 import type { SelectItem } from '~/types/components/select';
 import CommonBlockTitle from '~/components/common/blocks/CommonBlockTitle.vue';
+import CommonButton from '~/components/common/basic/CommonButton.vue';
+import { backupMapSettings } from '~/composables/settings';
+import { resetUserMapSettings } from '~/composables';
 
 const store = useStore();
+
+const resetActive = ref(false);
 
 // For type safety
 const countersOptions: Record<Required<IUserMapSettings['airportsCounters']>['departuresMode'], string> = {
     total: 'Total departures',
-    totalMoving: 'Total departures with GS > 0',
+    totalMoving: 'Total departures (moving only)',
     airborne: 'Airborne departures',
     ground: 'Departing (default)',
-    groundMoving: 'Departing with GS > 0',
+    groundMoving: 'Departing (moving only)',
     hide: 'Hide',
 };
 
 const countersArrivalOptions: Record<Required<IUserMapSettings['airportsCounters']>['departuresMode'], string> = {
     total: 'Total arrivals',
-    totalMoving: 'Total arrivals with GS > 0',
+    totalMoving: 'Total arrivals (moving only)',
     airborne: 'Airborne arrivals',
     ground: 'Landed (default)',
-    groundMoving: 'Landed with GS > 0',
+    groundMoving: 'Landed (moving only)',
     hide: 'Hide',
 };
 
@@ -125,7 +166,7 @@ const horizontalOptions: Record<Required<IUserMapSettings['airportsCounters']>['
     total: 'Total airport traffic',
     prefiles: 'Prefiles (default)',
     ground: 'Ground',
-    groundMoving: 'Ground with GS > 0',
+    groundMoving: 'Ground (moving only)',
     hide: 'Hide',
 };
 
@@ -133,4 +174,17 @@ const horizontalSelectOptions = Object.entries(horizontalOptions).map(([key, val
     text: value,
     value: key,
 } satisfies SelectItem));
+
+const scaleOptions = computed<SelectItem[]>(() => {
+    const options: SelectItem[] = [];
+
+    for (let i = 0.5; i <= 1.51; i += 0.05) {
+        options.unshift({
+            value: Number(i.toFixed(2)),
+            text: `x${ i.toFixed(2) }`,
+        });
+    }
+
+    return options;
+});
 </script>