+++ src/store/index.ts
@@ -8,7 +8,8 @@ import type { VatsimLiveData, VatsimLiveDataShort, VatsimMandatoryData } from '~
 import { setVatsimDataStore } from '~/composables/data';
 import { useMapStore } from '~/store/map';
 import type { Coordinate } from 'ol/coordinate';
-import type { UserMapSettings } from '~/utils/backend/map-settings';
+import type { UserMapPreset, UserMapSettings } from '~/utils/backend/map-settings';
+import type { TurnsBulkReturn } from '~/server/api/data/vatsim/pilot/turns';
 
 export interface SiteConfig {
     hideSectors?: boolean;
@@ -41,6 +42,7 @@ export const useStore = defineStore('index', {
         theme: 'default' as ThemesList,
         localSettings: {} as UserLocalSettings,
         mapSettings: {} as UserMapSettings,
+        mapPresets: [] as UserMapPreset[],
         config: {} as SiteConfig,
 
         showPilotStats: false,
@@ -51,6 +53,8 @@ export const useStore = defineStore('index', {
 
         updateRequired: false,
         isTabVisible: false,
+
+        loginPopup: false,
 
         viewport: {
             width: 0,
@@ -106,7 +110,7 @@ export const useStore = defineStore('index', {
 
                     if (!mapStore.localTurns.size) return;
 
-                    mapStore.turnsResponse = await $fetch('/api/data/vatsim/pilot/turns', {
+                    mapStore.turnsResponse = await $fetch<TurnsBulkReturn[]>('/api/data/vatsim/pilot/turns', {
                         method: 'POST',
                         body: [...mapStore.localTurns],
                     });